+++ test/devise/async_test.rb
@@ -11,14 +11,8 @@ module Devise
       Async.backend.must_equal :something
     end
 
-    it "stores mailer config" do
-      Async.mailer = "MyMailer"
-      Async.mailer.must_equal "MyMailer"
-    end
-
     after do
       Async.backend = :resque
-      Async.mailer = "Devise::Mailer"
     end
   end
 end